+++ Roman Bychkov/LearningSystem/LearningSystem.BLL/Services/CommentsService.cs
@@ -1,6 +1,6 @@
 ﻿namespace LearningSystem.BLL.Services
 {
-    public class CommentsService:ICommentsService
+    public class CommentsService : ICommentsService
     {
         private ICommentsRepository _context;
         public CommentsService(ICommentsRepository context)
@@ -9,11 +9,17 @@
         }
         public async Task AddAsync(Comment item)
         {
+            if (item is null)
+                throw new ArgumentNullException("comment");
+            if (string.IsNullOrWhiteSpace(item.Content) || item.Content.Length > 250 || item.Content.Length == 0)
+                throw new ArgumentException("Invalid length of comment");
             await _context.AddAsync(item);
         }
 
         public async Task DeleteAsync(Comment item)
         {
+            if (item is null)
+                throw new ArgumentNullException("comment");
             await _context.DeleteAsync(item);
         }
 
@@ -29,6 +35,10 @@
 
         public async Task UpdateAsync(Comment item)
         {
+            if (item is null)
+                throw new ArgumentNullException("comment");
+            if (string.IsNullOrWhiteSpace(item.Content) || item.Content.Length > 250 || item.Content.Length == 0)
+                throw new ArgumentException("Invalid length of comment");
             await _context.UpdateAsync(item);
         }
     }
