+++ Roman Bychkov/LearningSystem/LearningSystem.BLL/Services/CoursesService.cs
@@ -1,6 +1,6 @@
 ﻿namespace LearningSystem.BLL.Services
 {
-    public class CoursesService:ICoursesService
+    public class CoursesService : ICoursesService
     {
         private ICoursesRepository _context;
         public CoursesService(ICoursesRepository context)
@@ -9,11 +9,24 @@
         }
         public async Task AddAsync(Course item)
         {
+            if (item is null)
+                throw new ArgumentNullException("item");
+            if (string.IsNullOrWhiteSpace(item.CourseName) || item.CourseName.Length > 50)
+                throw new ArgumentException("Invalid Description");
+            if (string.IsNullOrWhiteSpace(item.Description) || item.Description.Length > 200)
+                throw new ArgumentException("Invalid Description");
+            if (string.IsNullOrWhiteSpace(item.Content) || item.Content.Length > 10000)
+                throw new ArgumentException("Invalid Content");
+
             await _context.AddAsync(item);
         }
 
         public async Task DeleteAsync(Course item)
         {
+            if (item is null)
+            {
+                throw new ArgumentNullException(nameof(item));
+            }
             await _context.DeleteAsync(item);
         }
 
@@ -29,6 +42,15 @@
 
         public async Task UpdateAsync(Course item)
         {
+            if (item is null)
+                throw new ArgumentNullException("item");
+            if (string.IsNullOrWhiteSpace(item.CourseName) || item.CourseName.Length > 50)
+                throw new ArgumentException("Invalid Description");
+            if (string.IsNullOrWhiteSpace(item.Description) || item.Description.Length > 200)
+                throw new ArgumentException("Invalid Description");
+            if (string.IsNullOrWhiteSpace(item.Content) || item.Content.Length > 10000)
+                throw new ArgumentException("Invalid Content");
+            
             await _context.UpdateAsync(item);
         }
     }
