+++ Roman Bychkov/LearningSystem/LearningSystem.BLL/Services/UsersService.cs
@@ -1,6 +1,7 @@
-﻿using Microsoft.AspNetCore.Cryptography.KeyDerivation;
+﻿using Microsoft.AspNetCore.Http;
 using System.Security.Cryptography;
 using System.Text;
+using System.Text.RegularExpressions;
 
 namespace LearningSystem.BLL.Services
 {
@@ -13,8 +14,16 @@ namespace LearningSystem.BLL.Services
         }
         public async Task AddAsync(User item)
         {
-
+            if (item is null)
+                throw new ArgumentNullException("item");
+            if (string.IsNullOrWhiteSpace(item.UserName) || item.UserName.Length > 50)
+                throw new ArgumentException("Invalid Username");
+            if (string.IsNullOrWhiteSpace(item.Password) || item.UserName.Length < 3 || item.UserName.Length > 50)
+                throw new ArgumentException("Invalid Password");
+            if (!Regex.IsMatch(item.Email, @"^[^@\s]+@[^@\s]+\.[^@\s]+$", RegexOptions.IgnoreCase))
+                throw new ArgumentException("Invalid Email");
             item.Password = SHA256Managed(item.Password);
+
             await _context.AddAsync(item);
         }
         private string SHA256Managed(string password)
@@ -26,6 +35,10 @@ namespace LearningSystem.BLL.Services
         }
         public async Task DeleteAsync(User item)
         {
+            if (item is null)
+            {
+                throw new ArgumentNullException(nameof(item));
+            }
             await _context.DeleteAsync(item);
         }
 
@@ -41,21 +54,64 @@ namespace LearningSystem.BLL.Services
 
         public async Task UpdateAsync(User item)
         {
+            if (item is null)
+                throw new ArgumentNullException("item");
+            if (string.IsNullOrWhiteSpace(item.UserName) || item.UserName.Length > 50)
+                throw new ArgumentException("Invalid Username");
+            if (string.IsNullOrWhiteSpace(item.Password) || item.UserName.Length < 3 || item.UserName.Length > 50)
+                throw new ArgumentException("Invalid Password");
+            if (!Regex.IsMatch(item.Email, @"^[^@\s]+@[^@\s]+\.[^@\s]+$", RegexOptions.IgnoreCase))
+                throw new ArgumentException("Invalid Email");
+            if (!Regex.IsMatch(item.Image, @"\b\w+\.(png|jpe?g)\b", RegexOptions.IgnoreCase))
+                throw new ArgumentException("Invalid Image");
+
             await _context.UpdateAsync(item);
         }
 
-        public async Task<User> GetValueByСonditionAsync(Func<User, string> valueSelector, string value)
+        public async Task<User> GetValueByСonditionAsync<T>(Func<User, T> valueSelector, T value)
         {
             return await _context.GetValueByСonditionAsync(valueSelector, value);
         }
 
         public async Task<User> GetUserByLoginPasswordAsync(string login, string password)
         {
-            if(!string.IsNullOrEmpty(password))
+            if (string.IsNullOrEmpty(login))
+                throw new ArgumentNullException(nameof(login));
+            if (string.IsNullOrEmpty(password))
+                throw new ArgumentNullException(nameof(password));
+
+            if (!string.IsNullOrEmpty(password))
                 password = SHA256Managed(password);
             return await _context.GetUserByLoginPasswordAsync(login, password);
         }
 
-       
+        public async Task<string> AddImage(string nameFolder, string name, IFormFile file)
+        {
+            if (string.IsNullOrWhiteSpace(nameFolder))
+                throw new ArgumentNullException(nameof(nameFolder));
+            if (string.IsNullOrWhiteSpace(name))
+                throw new ArgumentNullException(nameof(name));
+            if (file is null)
+                throw new ArgumentNullException(nameof(file));
+
+            string uploadPath = "wwwroot/" + nameFolder;
+
+            Directory.CreateDirectory(uploadPath);
+
+
+            var fileName = $"{name}.{Path.GetExtension(file.FileName).Replace(".", "")}";
+
+            string fullPath = $"{uploadPath}/{fileName}";
+
+
+            using (var fileStream = new FileStream(fullPath, FileMode.Create))
+            {
+                await file.CopyToAsync(fileStream);
+            }
+
+            await _context.AddImage("/" + nameFolder + "/" + fileName, name);
+
+            return "/" + nameFolder + "/" + fileName;
+        }
     }
 }
